requirements.txt
pycryptodome ==3.23.0